refactor/아이콘 파일 통합 및 관리 일원화
