--- src/shared/assets/SVGicons/24pxIcon.jsx
@@ -1,5 +1,0 @@
-export const CloseSVG = ({ stroke }) => (
-  <svg width='24' height='24' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'>
-    <path d='M20 4L4 20M20 20L4 4' stroke={stroke} strokeWidth='2' strokeLinecap='round' />
-  </svg>
-);
