--- src/shared/assets/SVGicons/32pxIcon.jsx
@@ -1,44 +1,0 @@
-export const FavoritesSVG = ({ fill, stroke }) => (
-  // stroke = #919191
-
-  <svg width='32' height='32' viewBox='0 0 32 32' fill={fill} xmlns='http://www.w3.org/2000/svg'>
-    <path
-      d='M11 7C8.239 7 6 9.216 6 11.95C6 14.157 6.875 19.395 15.488 24.69C15.6423 24.7839 15.8194 24.8335 16 24.8335C16.1806 24.8335 16.3577 24.7839 16.512 24.69C25.125 19.395 26 14.157 26 11.95C26 9.216 23.761 7 21 7C18.239 7 16 10 16 10C16 10 13.761 7 11 7Z'
-      stroke={stroke}
-      strokeWidth='1.5'
-      strokeLinecap='round'
-      strokeLinejoin='round'
-    />
-  </svg>
-);
-export const MoonSVG = ({ stroke }) => (
-  <svg width='32' height='32' viewBox='0 0 32 32' fill='none' xmlns='http://www.w3.org/2000/svg'>
-    <path
-      d='M28.7998 19.5194C27.5968 19.8843 26.3204 20.0804 24.9982 20.0804C17.7742 20.0804 11.918 14.2242 11.918 7.00023C11.918 5.67861 12.114 4.40277 12.4785 3.2002C7.10863 4.82871 3.19995 9.81759 3.19995 15.7194C3.19995 22.9434 9.05616 28.7996 16.2802 28.7996C22.1825 28.7996 27.1718 24.8902 28.7998 19.5194Z'
-      stroke={stroke}
-      strokeWidth='2'
-      strokeLinejoin='round'
-    />
-  </svg>
-);
-export const SunSVG = ({ stroke }) => (
-  <svg width='32' height='32' viewBox='0 0 32 32' fill='none' xmlns='http://www.w3.org/2000/svg'>
-    <path
-      d='M16 7.3513V3.19995M16 28.8V24.6486M24.6486 16H28.8M3.19995 16H7.3513M22.1159 9.88472L25.0513 6.94928M6.94829 25.0509L9.88373 22.1155M22.1159 22.1152L25.0513 25.0506M6.94829 6.94896L9.88373 9.88441M20.7583 15.8312C20.7583 18.4822 18.6093 20.6312 15.9583 20.6312C13.3073 20.6312 11.1583 18.4822 11.1583 15.8312C11.1583 13.1802 13.3073 11.0312 15.9583 11.0312C18.6093 11.0312 20.7583 13.1802 20.7583 15.8312Z'
-      stroke={stroke}
-      strokeWidth='2'
-      strokeLinecap='round'
-      strokeLinejoin='round'
-    />
-  </svg>
-);
-export const SearchSVG = ({ stroke }) => (
-  <svg width='32' height='32' viewBox='0 0 32 32' fill='none' xmlns='http://www.w3.org/2000/svg'>
-    <path
-      d='M22.5692 22.7198L27.2 27.1998M15.2 9.5998C17.851 9.5998 20 11.7488 20 14.3998M25.7067 15.2531C25.7067 21.0264 21.0266 25.7065 15.2534 25.7065C9.48017 25.7065 4.80005 21.0264 4.80005 15.2531C4.80005 9.47992 9.48017 4.7998 15.2534 4.7998C21.0266 4.7998 25.7067 9.47992 25.7067 15.2531Z'
-      stroke={stroke}
-      strokeWidth='2'
-      strokeLinecap='round'
-    />
-  </svg>
-);
