--- src/shared/assets/SVGicons/40pxIcon.jsx
@@ -1,8 +1,0 @@
-export const PlaySVG = ({ fill }) => (
-  <svg width='60' height='60' viewBox='0 0 40 40' fill='none' xmlns='http://www.w3.org/2000/svg'>
-    <path
-      d='M25.9991 21.3491L25.4535 20.511L25.4395 20.5201L25.4259 20.5296L25.9991 21.3491ZM25.9991 19.0554L25.4061 19.8607L25.4507 19.8935L25.4987 19.9212L25.9991 19.0554ZM16.8067 12.2866L17.3996 11.4814L17.3714 11.4606L17.3418 11.4418L16.8067 12.2866ZM14.8796 13.2976L15.8796 13.2947L15.8796 13.286L15.8794 13.2772L14.8796 13.2976ZM14.9181 26.5938L13.9181 26.5967L13.9182 26.625L13.9199 26.6532L14.9181 26.5938ZM16.9159 27.7018L17.4001 28.5768L17.446 28.5514L17.489 28.5213L16.9159 27.7018ZM10 4V5H30V4V3H10V4ZM36 10H35V30H36H37V10H36ZM30 36V35H10V36V37H30V36ZM4 30H5V10H4H3V30H4ZM10 36V35C7.23858 35 5 32.7614 5 30H4H3C3 33.866 6.13401 37 10 37V36ZM36 30H35C35 32.7614 32.7614 35 30 35V36V37C33.866 37 37 33.866 37 30H36ZM30 4V5C32.7614 5 35 7.23858 35 10H36H37C37 6.13401 33.866 3 30 3V4ZM10 4V3C6.13401 3 3 6.13401 3 10H4H5C5 7.23858 7.23858 5 10 5V4ZM25.9991 21.3491L26.5446 22.1871C27.2141 21.7513 27.8804 21.0852 27.8791 20.1677C27.8778 19.2278 27.183 18.5847 26.4994 18.1896L25.9991 19.0554L25.4987 19.9212C25.7205 20.0494 25.8277 20.1529 25.8718 20.2091C25.9089 20.2564 25.8792 20.2392 25.8791 20.1705C25.879 20.1054 25.9059 20.1006 25.856 20.168C25.8016 20.2415 25.6828 20.3617 25.4535 20.511L25.9991 21.3491ZM25.9991 19.0554L26.592 18.2502L17.3996 11.4814L16.8067 12.2866L16.2138 13.0919L25.4061 19.8607L25.9991 19.0554ZM16.8067 12.2866L17.3418 11.4418C16.6806 11.023 15.8245 10.8274 15.0531 11.1876C14.2265 11.5735 13.8615 12.4192 13.8798 13.318L14.8796 13.2976L15.8794 13.2772C15.8756 13.0928 15.9109 13.0084 15.9221 12.9868C15.9295 12.9728 15.9227 12.9888 15.8993 12.9998C15.8597 13.0183 15.9799 12.9466 16.2716 13.1314L16.8067 12.2866ZM14.8796 13.2976L13.8796 13.3005L13.9181 26.5967L14.9181 26.5938L15.9181 26.5909L15.8796 13.2947L14.8796 13.2976ZM14.9181 26.5938L13.9199 26.6532C13.968 27.4609 14.2394 28.3471 15.0485 28.7793C15.8472 29.2058 16.7261 28.9498 17.4001 28.5768L16.9159 27.7018L16.4317 26.8269C16.2184 26.9449 16.0745 26.986 15.9964 26.9968C15.9263 27.0065 15.9414 26.9888 15.9907 27.0151C16.0416 27.0423 16.0376 27.0676 16.0094 27.0076C15.977 26.9384 15.9317 26.7923 15.9164 26.5344L14.9181 26.5938ZM16.9159 27.7018L17.489 28.5213L26.5722 22.1685L25.9991 21.3491L25.4259 20.5296L16.3428 26.8824L16.9159 27.7018Z'
-      fill={fill}
-    />
-  </svg>
-);
