+++ src/shared/assets/SVGicons/index.jsx
@@ -1,0 +1,65 @@
+export const PlaySVG = ({ size, fill }) => (
+  <svg width={size} height={size} viewBox='0 0 40 40' fill='none' xmlns='http://www.w3.org/2000/svg'>
+    <path
+      d='M25.9991 21.3491L25.4535 20.511L25.4395 20.5201L25.4259 20.5296L25.9991 21.3491ZM25.9991 19.0554L25.4061 19.8607L25.4507 19.8935L25.4987 19.9212L25.9991 19.0554ZM16.8067 12.2866L17.3996 11.4814L17.3714 11.4606L17.3418 11.4418L16.8067 12.2866ZM14.8796 13.2976L15.8796 13.2947L15.8796 13.286L15.8794 13.2772L14.8796 13.2976ZM14.9181 26.5938L13.9181 26.5967L13.9182 26.625L13.9199 26.6532L14.9181 26.5938ZM16.9159 27.7018L17.4001 28.5768L17.446 28.5514L17.489 28.5213L16.9159 27.7018ZM10 4V5H30V4V3H10V4ZM36 10H35V30H36H37V10H36ZM30 36V35H10V36V37H30V36ZM4 30H5V10H4H3V30H4ZM10 36V35C7.23858 35 5 32.7614 5 30H4H3C3 33.866 6.13401 37 10 37V36ZM36 30H35C35 32.7614 32.7614 35 30 35V36V37C33.866 37 37 33.866 37 30H36ZM30 4V5C32.7614 5 35 7.23858 35 10H36H37C37 6.13401 33.866 3 30 3V4ZM10 4V3C6.13401 3 3 6.13401 3 10H4H5C5 7.23858 7.23858 5 10 5V4ZM25.9991 21.3491L26.5446 22.1871C27.2141 21.7513 27.8804 21.0852 27.8791 20.1677C27.8778 19.2278 27.183 18.5847 26.4994 18.1896L25.9991 19.0554L25.4987 19.9212C25.7205 20.0494 25.8277 20.1529 25.8718 20.2091C25.9089 20.2564 25.8792 20.2392 25.8791 20.1705C25.879 20.1054 25.9059 20.1006 25.856 20.168C25.8016 20.2415 25.6828 20.3617 25.4535 20.511L25.9991 21.3491ZM25.9991 19.0554L26.592 18.2502L17.3996 11.4814L16.8067 12.2866L16.2138 13.0919L25.4061 19.8607L25.9991 19.0554ZM16.8067 12.2866L17.3418 11.4418C16.6806 11.023 15.8245 10.8274 15.0531 11.1876C14.2265 11.5735 13.8615 12.4192 13.8798 13.318L14.8796 13.2976L15.8794 13.2772C15.8756 13.0928 15.9109 13.0084 15.9221 12.9868C15.9295 12.9728 15.9227 12.9888 15.8993 12.9998C15.8597 13.0183 15.9799 12.9466 16.2716 13.1314L16.8067 12.2866ZM14.8796 13.2976L13.8796 13.3005L13.9181 26.5967L14.9181 26.5938L15.9181 26.5909L15.8796 13.2947L14.8796 13.2976ZM14.9181 26.5938L13.9199 26.6532C13.968 27.4609 14.2394 28.3471 15.0485 28.7793C15.8472 29.2058 16.7261 28.9498 17.4001 28.5768L16.9159 27.7018L16.4317 26.8269C16.2184 26.9449 16.0745 26.986 15.9964 26.9968C15.9263 27.0065 15.9414 26.9888 15.9907 27.0151C16.0416 27.0423 16.0376 27.0676 16.0094 27.0076C15.977 26.9384 15.9317 26.7923 15.9164 26.5344L14.9181 26.5938ZM16.9159 27.7018L17.489 28.5213L26.5722 22.1685L25.9991 21.3491L25.4259 20.5296L16.3428 26.8824L16.9159 27.7018Z'
+      fill={fill}
+    />
+  </svg>
+);
+
+export const FavoritesSVG = ({ size, fill, stroke }) => (
+  // stroke = #919191
+
+  <svg width={size} height={size} viewBox='0 0 32 32' fill={fill} xmlns='http://www.w3.org/2000/svg'>
+    <path
+      d='M11 7C8.239 7 6 9.216 6 11.95C6 14.157 6.875 19.395 15.488 24.69C15.6423 24.7839 15.8194 24.8335 16 24.8335C16.1806 24.8335 16.3577 24.7839 16.512 24.69C25.125 19.395 26 14.157 26 11.95C26 9.216 23.761 7 21 7C18.239 7 16 10 16 10C16 10 13.761 7 11 7Z'
+      stroke={stroke}
+      strokeWidth='1.5'
+      strokeLinecap='round'
+      strokeLinejoin='round'
+    />
+  </svg>
+);
+
+export const MoonSVG = ({ size, stroke }) => (
+  <svg width={size} height={size} viewBox='0 0 32 32' fill='none' xmlns='http://www.w3.org/2000/svg'>
+    <path
+      d='M28.7998 19.5194C27.5968 19.8843 26.3204 20.0804 24.9982 20.0804C17.7742 20.0804 11.918 14.2242 11.918 7.00023C11.918 5.67861 12.114 4.40277 12.4785 3.2002C7.10863 4.82871 3.19995 9.81759 3.19995 15.7194C3.19995 22.9434 9.05616 28.7996 16.2802 28.7996C22.1825 28.7996 27.1718 24.8902 28.7998 19.5194Z'
+      stroke={stroke}
+      strokeWidth='2'
+      strokeLinejoin='round'
+    />
+  </svg>
+);
+export const SunSVG = ({ size, stroke }) => (
+  <svg width={size} height={size} viewBox='0 0 32 32' fill='none' xmlns='http://www.w3.org/2000/svg'>
+    <path
+      d='M16 7.3513V3.19995M16 28.8V24.6486M24.6486 16H28.8M3.19995 16H7.3513M22.1159 9.88472L25.0513 6.94928M6.94829 25.0509L9.88373 22.1155M22.1159 22.1152L25.0513 25.0506M6.94829 6.94896L9.88373 9.88441M20.7583 15.8312C20.7583 18.4822 18.6093 20.6312 15.9583 20.6312C13.3073 20.6312 11.1583 18.4822 11.1583 15.8312C11.1583 13.1802 13.3073 11.0312 15.9583 11.0312C18.6093 11.0312 20.7583 13.1802 20.7583 15.8312Z'
+      stroke={stroke}
+      strokeWidth='2'
+      strokeLinecap='round'
+      strokeLinejoin='round'
+    />
+  </svg>
+);
+export const SearchSVG = ({ size, stroke }) => (
+  <svg width={size} height={size} viewBox='0 0 32 32' fill='none' xmlns='http://www.w3.org/2000/svg'>
+    <path
+      d='M22.5692 22.7198L27.2 27.1998M15.2 9.5998C17.851 9.5998 20 11.7488 20 14.3998M25.7067 15.2531C25.7067 21.0264 21.0266 25.7065 15.2534 25.7065C9.48017 25.7065 4.80005 21.0264 4.80005 15.2531C4.80005 9.47992 9.48017 4.7998 15.2534 4.7998C21.0266 4.7998 25.7067 9.47992 25.7067 15.2531Z'
+      stroke={stroke}
+      strokeWidth='2'
+      strokeLinecap='round'
+    />
+  </svg>
+);
+
+export const CloseSVG = ({ size, stroke }) => (
+  <svg width={size} height={size} viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'>
+    <path d='M20 4L4 20M20 20L4 4' stroke={stroke} strokeWidth='2' strokeLinecap='round' />
+  </svg>
+);
+export const MenuSVG = ({ size, stroke }) => (
+  <svg width={size} height={size} viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'>
+    <path d='M20 18H4M20 12H4M20 6H4' stroke={stroke} stroke-width='2' stroke-linecap='round' />
+  </svg>
+);
