+++ src/shared/layouts/Navbar/index.jsx
@@ -1,11 +1,9 @@
-import { MoonSVG, SearchSVG } from '@/shared/assets/SVGicons/32pxIcon';
-import { PlaySVG } from '@/shared/assets/SVGicons/40pxIcon';
 import { useMovieList } from '@/api/movieHooks';
 import * as S from './Navbar.styled';
 import { useDispatch, useSelector } from 'react-redux';
 import { searchModeSlice, searchQuerySlice } from '@/app/store';
-import { CloseSVG } from '@/shared/assets/SVGicons/24pxIcon';
 import SearchMovieCard from '@/shared/ui/Card/SearchMovieCard';
+import { MenuSVG, MoonSVG, PlaySVG, SearchSVG, CloseSVG } from '@/shared/assets/SVGicons';
 
 const NavBar = () => {
   const dispatch = useDispatch();
@@ -23,19 +21,22 @@ const NavBar = () => {
       <S.MainBox>
         <S.FirstBox>
           <S.Logo to={'./'} onClick={() => dispatch(searchModeSlice.actions.close())}>
-            {PlaySVG({ fill: '#000' })}
+            {PlaySVG({ fill: '#000', size: 60 })}
           </S.Logo>
 
           <S.BtBox>
             <S.Bt>로그인</S.Bt>
             <S.Bt>회원가입</S.Bt>
-            <S.Icon>{MoonSVG({ stroke: '#000' })}</S.Icon>
+            <S.Icon>{MoonSVG({ size: '32', stroke: '#000' })}</S.Icon>
             {toggle ? (
-              <S.Icon onClick={() => ToggleAndClear()}>{CloseSVG({ stroke: '#000' })}</S.Icon>
+              <S.Icon onClick={() => ToggleAndClear()}>{CloseSVG({ size: '32', stroke: '#000' })}</S.Icon>
             ) : (
-              <S.Icon onClick={() => dispatch(searchModeSlice.actions.open())}>{SearchSVG({ stroke: '#000' })}</S.Icon>
+              <S.Icon onClick={() => dispatch(searchModeSlice.actions.open())}>{SearchSVG({ size: '32', stroke: '#000' })}</S.Icon>
             )}
           </S.BtBox>
+          <S.Menu>
+            <S.Icon>{MenuSVG({ size: '32', stroke: '#000' })}</S.Icon>
+          </S.Menu>
         </S.FirstBox>
       </S.MainBox>
 
@@ -43,7 +44,7 @@ const NavBar = () => {
         <>
           <S.MainBox>
             <S.InputBox>
-              <S.Icon style={{ marginRight: '10px' }}>{SearchSVG({ stroke: '#999' })}</S.Icon>
+              <S.Icon style={{ marginRight: '10px' }}>{SearchSVG({ size: '32', stroke: '#999' })}</S.Icon>
               <S.Input placeholder='제목를 입력하세요.' onChange={(e) => dispatch(searchQuerySlice.actions.setQuery(e.target.value))} />
             </S.InputBox>
           </S.MainBox>
